refactor JsonUpload
